ci: split out engine module
