+++ .dagger/sdk_rust.go
@@ -62,11 +62,8 @@ func (r RustSDK) Lint(ctx context.Context) error {
 
 // Test the Rust SDK
 func (r RustSDK) Test(ctx context.Context) error {
-	installer, err := r.Dagger.installer(ctx, "sdk")
-	if err != nil {
-		return err
-	}
-	_, err = r.rustBase(rustDockerStable).
+	installer := r.Dagger.installer("sdk")
+	_, err := r.rustBase(rustDockerStable).
 		With(installer).
 		WithExec([]string{"rustc", "--version"}).
 		WithExec([]string{"cargo", "test", "--release", "--all"}).
@@ -76,11 +73,7 @@ func (r RustSDK) Test(ctx context.Context) error {
 
 // Regenerate the Rust SDK API
 func (r RustSDK) Generate(ctx context.Context) (*dagger.Directory, error) {
-	installer, err := r.Dagger.installer(ctx, "sdk")
-	if err != nil {
-		return nil, err
-	}
-
+	installer := r.Dagger.installer("sdk")
 	generated := r.rustBase(rustDockerStable).
 		With(installer).
 		WithExec([]string{"cargo", "run", "-p", "dagger-bootstrap", "generate", "--output", fmt.Sprintf("/%s", rustGeneratedAPIPath)}).
