+++ cmd/engine/.dagger/go.mod
@@ -1,33 +1,40 @@
-module github.com/dagger/dagger/.dagger
+module github.com/dagger/dagger/cmd/engine/.dagger
 
-go 1.23.2
+go 1.23.8
 
-require github.com/dagger/dagger/engine/distconsts v0.18.6
+require (
+	github.com/dagger/dagger/engine/distconsts v0.18.5
+	github.com/dagger/dagger/sdk/typescript/runtime v0.15.3
+)
 
 replace (
-	github.com/dagger/dagger/engine/distconsts => ../engine/distconsts
-	github.com/dagger/dagger/sdk/typescript/runtime => ../sdk/typescript/runtime
+	github.com/dagger/dagger/engine/distconsts => ../../../engine/distconsts
+	github.com/dagger/dagger/sdk/typescript/runtime => ../../../sdk/typescript/runtime
 )
 
 require (
 	github.com/99designs/gqlgen v0.17.70
-	github.com/BurntSushi/toml v0.3.1
 	github.com/Khan/genqlient v0.8.0
-	github.com/containerd/platforms v0.2.1
-	github.com/moby/buildkit v0.14.0-rc1.0.20240603193914-3d789eb740a9
+	github.com/containerd/platforms v1.0.0-rc.1
+	github.com/moby/buildkit v0.21.0
+	github.com/opencontainers/image-spec v1.1.1
 	github.com/vektah/gqlparser/v2 v2.5.23
 	go.opentelemetry.io/otel v1.34.0
 	go.opentelemetry.io/otel/exporters/otlp/otlplog/otlploggrpc v0.8.0
 	go.opentelemetry.io/otel/exporters/otlp/otlplog/otlploghttp v0.8.0
+	go.opentelemetry.io/otel/exporters/otlp/otlpmetric/otlpmetricgrpc v1.32.0
+	go.opentelemetry.io/otel/exporters/otlp/otlpmetric/otlpmetrichttp v1.32.0
 	go.opentelemetry.io/otel/exporters/otlp/otlptrace/otlptracegrpc v1.32.0
 	go.opentelemetry.io/otel/exporters/otlp/otlptrace/otlptracehttp v1.32.0
 	go.opentelemetry.io/otel/log v0.8.0
+	go.opentelemetry.io/otel/metric v1.34.0
 	go.opentelemetry.io/otel/sdk v1.34.0
 	go.opentelemetry.io/otel/sdk/log v0.8.0
+	go.opentelemetry.io/otel/sdk/metric v1.34.0
 	go.opentelemetry.io/otel/trace v1.34.0
 	go.opentelemetry.io/proto/otlp v1.3.1
-	golang.org/x/mod v0.24.0
-	golang.org/x/sync v0.12.0
+	golang.org/x/exp v0.0.0-20250408133849-7e4ce0ab07d0
+	golang.org/x/sync v0.13.0
 	google.golang.org/grpc v1.71.0
 )
 
@@ -39,20 +46,13 @@ require (
 	github.com/google/uuid v1.6.0 // indirect
 	github.com/grpc-ecosystem/grpc-gateway/v2 v2.23.0 // indirect
 	github.com/opencontainers/go-digest v1.0.0 // indirect
-	github.com/opencontainers/image-spec v1.1.0 // indirect
-	github.com/pkg/errors v0.9.1 // indirect
-	github.com/sergi/go-diff v1.3.2-0.20230802210424-5b0b94c5c0d3 // indirect
 	github.com/sirupsen/logrus v1.9.3 // indirect
 	github.com/sosodev/duration v1.3.1 // indirect
 	go.opentelemetry.io/auto/sdk v1.1.0 // indirect
-	go.opentelemetry.io/otel/exporters/otlp/otlpmetric/otlpmetricgrpc v1.32.0
-	go.opentelemetry.io/otel/exporters/otlp/otlpmetric/otlpmetrichttp v1.32.0
 	go.opentelemetry.io/otel/exporters/otlp/otlptrace v1.32.0 // indirect
-	go.opentelemetry.io/otel/metric v1.34.0
-	go.opentelemetry.io/otel/sdk/metric v1.34.0
-	golang.org/x/net v0.38.0 // indirect
-	golang.org/x/sys v0.31.0 // indirect
-	golang.org/x/text v0.23.0 // indirect
+	golang.org/x/net v0.39.0 // indirect
+	golang.org/x/sys v0.32.0 // indirect
+	golang.org/x/text v0.24.0 // indirect
 	google.golang.org/genproto/googleapis/api v0.0.0-20250106144421-5f5ef82da422 // indirect
 	google.golang.org/genproto/googleapis/rpc v0.0.0-20250115164207-1a7da9e5054f // indirect
 	google.golang.org/protobuf v1.36.6 // indirect
