+++ modules/compatcheck/dagger.json
@@ -1,6 +1,6 @@
 {
   "name": "compatcheck",
-  "engineVersion": "v0.12.5",
+  "engineVersion": "v0.18.5",
   "sdk": {
     "source": "go"
   },
